Adds ProcessorPipeline to Aztec. Adds preProcessor and postProcessor to Aztec.TextView.
